+++ samples/kmp/build.gradle.kts
@@ -1,7 +1,9 @@
 plugins {
-    id("com.android.application") version "8.7.0" apply false
-    id("org.jetbrains.kotlin.android") version "2.3.10" apply false
-    id("org.jetbrains.kotlin.plugin.compose") version "2.3.10" apply false
-    id("com.android.library") version "8.7.0" apply false
+    id("com.android.application") version "9.2.1" apply false
+    id("org.jetbrains.kotlin.android") version "2.4.20" apply false
+    id("org.jetbrains.kotlin.multiplatform") version "2.4.20" apply false
+    id("org.jetbrains.kotlin.plugin.compose") version "2.4.20" apply false
+    id("org.jetbrains.compose") version "1.12.0" apply false
+    id("com.android.library") version "9.2.1" apply false
     id("org.jlleitschuh.gradle.ktlint") version "12.1.2" apply false
 }
