+++ samples/kmp/gradle.properties
@@ -2,3 +2,5 @@ org.gradle.jvmargs=-Xmx2048m -Dfile.encoding=UTF-8
 android.useAndroidX=true
 kotlin.code.style=official
 android.nonTransitiveRClass=true
+android.builtInKotlin=false
+android.newDsl=false
