+++ samples/kmp/shared/src/commonMain/kotlin/jp/jig/glasses/sample/kmp/ui/LauncherScreen.kt
@@ -1,0 +1,316 @@
+package jp.jig.glasses.sample.kmp.ui
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.HorizontalDivider
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedButton
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.DisposableEffect
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import app.jigglass.glass.CommandManager
+import app.jigglass.glass.GestureType
+import app.jigglass.glass.GlassClient
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.filterNotNull
+import kotlinx.coroutines.flow.first
+import kotlinx.coroutines.isActive
+import kotlinx.coroutines.launch
+import kotlin.math.abs
+import kotlin.math.roundToInt
+
+/** アイコンの一辺。96角6枚でもグラスの画像バッファに収まる */
+private const val ICON_SIZE = 96
+private const val ICON_GAP = 48
+private const val ICON_ROW_GAP = 16
+private const val ICON_TOP_Y = 16
+
+/** 3列2段。画像は id 0..7 なので6枚まで置ける */
+private const val GRID_COLUMNS = 3
+private const val GRID_ROWS = 2
+
+/** 説明テキストの矩形。アイコンの下の余りに置く */
+private const val TEXT_X = 48
+private const val TEXT_WIDTH = 480
+private const val TEXT_HEIGHT = 112
+
+private const val CANVAS_WIDTH = 576
+
+/** アイコンを横に並べたときの左端 */
+private const val GRID_LEFT = (CANVAS_WIDTH - (GRID_COLUMNS * ICON_SIZE + (GRID_COLUMNS - 1) * ICON_GAP)) / 2
+
+/** 隣の列へ移るまでの振り向き量 */
+private const val STEP_DEGREES = 20f
+
+/** 上の段へ移るまでの見上げ量。ピッチは上向きが負 */
+private const val ROW_STEP_DEGREES = 15f
+
+/** 首の向きそのものを差すポインター。テキストは画像の手前に描かれる */
+private const val POINTER_TEXT = "・"
+private const val POINTER_ELEMENT_ID = 1
+private const val POINTER_SIZE = 32
+private const val POINTER_MOVE_PX = 8
+
+/** 境目での行き来を防ぐ遊び。列はセル幅に対する割合、段は度 */
+private const val COLUMN_HYSTERESIS = 0.25f
+private const val ROW_HYSTERESIS_DEGREES = 7f
+
+/** 起点は下段の中央。楽な姿勢のまま選べる位置に置く */
+private const val DEFAULT_INDEX = (GRID_ROWS - 1) * GRID_COLUMNS + GRID_COLUMNS / 2
+
+/** ヨーとピッチを見に行く間隔。変わったときだけ送るので短くても送信は増えない */
+private const val CURSOR_POLL_MS = 100L
+
+/** ランチャーに並べる機能。アイコンは絵柄の代わりに漢字1字で描く */
+enum class LauncherItem(val label: String, val glyph: String, val description: String) {
+    KAWAKUDARI("かわくだり", "川", "首を振って岩をよけるゲーム。タップでノーマル、長押しでハード。"),
+    TELEPROMPTER("テレプロンプター", "読", "原稿を送って、行を進めながら読み上げる。"),
+    TRANSLATE("翻訳", "訳", "話した言葉をその場で訳して出す。"),
+    AI_CHAT("AI アシスタント", "問", "聞いたことに答えを返す。"),
+    IMAGE("画像を送る", "絵", "選んだ写真をグレースケールにして出す。"),
+    IMU("6DoF を見る", "角", "加速度・角速度・ピッチ・ヨーの生の値を並べる。"),
+}
+
+/**
+ * 6DoF でカーソルを動かすランチャー。
+ * アイコンはキャンバスの画像、説明はテキスト要素で、選択はグラスのシングルタップ。
+ */
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun LauncherScreen(client: GlassClient, onSelect: (LauncherItem) -> Unit, onBack: () -> Unit) {
+    val scope = rememberCoroutineScope()
+    val commandManager = remember(client) { client.createCommandManager() }
+    val imu = remember { MutableStateFlow<CommandManager.ImuData?>(null) }
+
+    val items = remember { LauncherItem.entries }
+    // フォーカスの有無で差し替えるので、2状態とも先に作っておく
+    val icons = remember { items.map { iconImage(it.glyph, false) to iconImage(it.glyph, true) } }
+
+    var focused by remember { mutableStateOf(DEFAULT_INDEX) }
+
+    DisposableEffect(commandManager) {
+        val jobs = listOf(
+            scope.launch {
+                commandManager.imuData.collect { imu.value = it }
+            },
+            scope.launch {
+                commandManager.gestureEvents.collect { gesture ->
+                    if (gesture == GestureType.SINGLE_TAP) onSelect(items[focused])
+                }
+            },
+        )
+        commandManager.startImuData()
+        onDispose {
+            jobs.forEach { it.cancel() }
+            commandManager.stopImuData()
+            commandManager.closeCanvas()
+        }
+    }
+
+    LaunchedEffect(Unit) {
+        // 画面に入った向きを下段の中央に合わせる
+        val center = imu.filterNotNull().first()
+
+        // 回り込んでも連続させるため、列は範囲を切らないセル番号で持つ
+        var columnCell = GRID_COLUMNS / 2
+        var pointerX = pointerX(columnCell.toFloat())
+
+        commandManager.sendCanvas(
+            listOf(
+                descriptionElement(items[focused]),
+                pointerElement(pointerX, focused / GRID_COLUMNS),
+            ),
+        )
+        items.indices.forEach { index ->
+            commandManager.sendCanvasImage(
+                id = index,
+                x = iconX(index),
+                y = iconY(index),
+                width = ICON_SIZE,
+                height = ICON_SIZE,
+                grayscale = icons[index].let { if (index == focused) it.second else it.first }.pixels,
+            )
+        }
+
+        while (isActive) {
+            delay(CURSOR_POLL_MS)
+            val data = imu.value ?: center
+            columnCell = columnCellFor(data, center, columnCell)
+            val row = rowFor(data, center, focused / GRID_COLUMNS)
+            val next = row * GRID_COLUMNS + columnCell.mod(GRID_COLUMNS)
+
+            // ポインターは離散したフォーカスではなく、首の向きのまま動かす
+            val nextPointerX = pointerX(columnPosition(data, center))
+            if (next == focused) {
+                if (abs(nextPointerX - pointerX) < POINTER_MOVE_PX) continue
+                pointerX = nextPointerX
+                commandManager.sendCanvasElements(listOf(pointerElement(pointerX, row)))
+                continue
+            }
+
+            val previous = focused
+            focused = next
+            pointerX = nextPointerX
+            commandManager.sendCanvasImage(
+                id = previous,
+                x = iconX(previous),
+                y = iconY(previous),
+                width = ICON_SIZE,
+                height = ICON_SIZE,
+                grayscale = icons[previous].first.pixels,
+            )
+            commandManager.sendCanvasImage(
+                id = next,
+                x = iconX(next),
+                y = iconY(next),
+                width = ICON_SIZE,
+                height = ICON_SIZE,
+                grayscale = icons[next].second.pixels,
+            )
+            commandManager.sendCanvasElements(
+                listOf(
+                    descriptionElement(items[next]),
+                    pointerElement(pointerX, row),
+                ),
+            )
+        }
+    }
+
+    Scaffold(
+        topBar = { TopAppBar(title = { Text("ランチャー") }) },
+    ) { padding ->
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(padding)
+                .padding(horizontal = 24.dp)
+                .verticalScroll(rememberScrollState()),
+        ) {
+            Text(
+                "首を振ってカーソルを動かし、グラスのシングルタップで選ぶ。" +
+                    "隣の列まで${STEP_DEGREES.toInt()}度、上の段は${ROW_STEP_DEGREES.toInt()}度見上げる。",
+                style = MaterialTheme.typography.bodySmall,
+            )
+            Spacer(Modifier.height(16.dp))
+
+            (0 until GRID_ROWS).forEach { row ->
+                Row(
+                    modifier = Modifier.fillMaxWidth(),
+                    horizontalArrangement = Arrangement.SpaceEvenly,
+                    verticalAlignment = Alignment.CenterVertically,
+                ) {
+                    (0 until GRID_COLUMNS).forEach { column ->
+                        val index = row * GRID_COLUMNS + column
+                        val item = items.getOrNull(index) ?: return@forEach
+                        val icon = icons[index].let { if (index == focused) it.second else it.first }
+                        Image(
+                            bitmap = icon.toImageBitmap(),
+                            contentDescription = item.label,
+                            modifier = Modifier.size(72.dp),
+                        )
+                    }
+                }
+                Spacer(Modifier.height(8.dp))
+            }
+
+            Spacer(Modifier.height(8.dp))
+            Text(items[focused].label, style = MaterialTheme.typography.titleMedium)
+            Text(items[focused].description, style = MaterialTheme.typography.bodyMedium)
+
+            HorizontalDivider(Modifier.padding(vertical = 16.dp))
+
+            OutlinedButton(
+                onClick = { onSelect(items[focused]) },
+                modifier = Modifier.fillMaxWidth(),
+            ) {
+                Text("${items[focused].label}をひらく")
+            }
+            Spacer(Modifier.height(8.dp))
+            OutlinedButton(onClick = onBack, modifier = Modifier.fillMaxWidth()) {
+                Text("戻る")
+            }
+            Spacer(Modifier.height(24.dp))
+        }
+    }
+}
+
+private fun descriptionElement(item: LauncherItem) = CommandManager.CanvasElement(
+    id = 0,
+    x = TEXT_X,
+    y = ICON_TOP_Y + GRID_ROWS * (ICON_SIZE + ICON_ROW_GAP),
+    width = TEXT_WIDTH,
+    height = TEXT_HEIGHT,
+    text = "${item.label}\n${item.description}",
+)
+
+private fun iconX(index: Int): Int = GRID_LEFT + (index % GRID_COLUMNS) * (ICON_SIZE + ICON_GAP)
+
+private fun iconY(index: Int): Int =
+    ICON_TOP_Y + (index / GRID_COLUMNS) * (ICON_SIZE + ICON_ROW_GAP)
+
+/** 首の向きをセル単位の連続値にする。回り込むぶん範囲は切らない */
+private fun columnPosition(current: CommandManager.ImuData, center: CommandManager.ImuData): Float =
+    GRID_COLUMNS / 2f - normalizeDegrees(current.yawDegrees - center.yawDegrees) / STEP_DEGREES
+
+private fun columnCellFor(
+    current: CommandManager.ImuData,
+    center: CommandManager.ImuData,
+    currentCell: Int,
+): Int {
+    val position = columnPosition(current, center)
+    // 半セルを少し越えるまで居座らせる。境目で振れても隣に落ちない
+    return if (abs(position - currentCell) > 0.5f + COLUMN_HYSTERESIS) position.roundToInt() else currentCell
+}
+
+private fun pointerX(position: Float): Int {
+    val cell = position.mod(GRID_COLUMNS.toFloat())
+    return (GRID_LEFT + cell * (ICON_SIZE + ICON_GAP) + (ICON_SIZE - POINTER_SIZE) / 2).roundToInt()
+}
+
+private fun pointerElement(x: Int, row: Int) = CommandManager.CanvasElement(
+    id = POINTER_ELEMENT_ID,
+    x = x,
+    y = ICON_TOP_Y + row * (ICON_SIZE + ICON_ROW_GAP),
+    width = POINTER_SIZE,
+    height = POINTER_SIZE,
+    text = POINTER_TEXT,
+)
+
+private fun rowFor(current: CommandManager.ImuData, center: CommandManager.ImuData, currentRow: Int): Int {
+    val delta = current.pitchDegrees - center.pitchDegrees
+    return when {
+        delta < -ROW_STEP_DEGREES -> 0
+        delta > -ROW_STEP_DEGREES + ROW_HYSTERESIS_DEGREES -> GRID_ROWS - 1
+        else -> currentRow
+    }
+}
+
+/**
+ * アイコンを2値で描く。グラスは3bitに量子化するので、中間の階調は残さない。
+ * フォーカスは枠の太さと白黒反転で、絵柄そのものに含める。
+ */
+private fun iconImage(glyph: String, focused: Boolean, size: Int = ICON_SIZE): GrayscaleImage =
+    renderLauncherIcon(glyph, focused, size)
